sales_data_etl/param/requirements.txt
mysql-connector-python
google-cloud-secret-manager